Move StreamBufferWriter to a separate file.
