+++ src/managed/Microsoft.Extensions.DependencyModel/DependencyContextWriter.Netstandard2_0.cs
@@ -2,7 +2,6 @@
 // Licensed under the MIT license. See LICENSE file in the project root for full license information.
 
 using System;
-using System.Buffers;
 using System.Collections.Generic;
 using System.Diagnostics;
 using System.IO;
@@ -11,77 +10,6 @@ using System.Text.Json;
 
 namespace Microsoft.Extensions.DependencyModel
 {
-    internal struct StreamBufferWriter : IBufferWriter<byte>, IDisposable
-    {
-        Stream _stream;
-        byte[] _rentedBuffer;
-
-        public StreamBufferWriter(Stream stream, int bufferSize = 256)
-        {
-            Debug.Assert(stream != null);
-
-            if (bufferSize <= 0)
-            {
-                throw new ArgumentNullException(nameof(bufferSize));
-            }
-
-            _rentedBuffer = ArrayPool<byte>.Shared.Rent(bufferSize);
-            _stream = stream;
-        }
-
-        public void Advance(int count)
-        {
-            _stream.Write(_rentedBuffer, 0, count);
-            _rentedBuffer.AsSpan(0, count).Clear();
-        }
-
-        /// <summary>
-        /// Returns rented buffer back to the pool
-        /// </summary>
-        public void Dispose()
-        {
-            _rentedBuffer.AsSpan().Clear();
-            ArrayPool<byte>.Shared.Return(_rentedBuffer);
-            _rentedBuffer = null;
-        }
-
-        public Memory<byte> GetMemory(int sizeHint = 0)
-        {
-            if (sizeHint > _rentedBuffer.Length)
-            {
-                var newSize = _rentedBuffer.Length * 2;
-                if (sizeHint != 0)
-                {
-                    newSize = sizeHint;
-                }
-                var temp = _rentedBuffer;
-                _rentedBuffer = ArrayPool<byte>.Shared.Rent(newSize);
-                temp.AsSpan().Clear();
-                ArrayPool<byte>.Shared.Return(temp);
-            }
-            Debug.Assert(_rentedBuffer.Length > 0);
-            return _rentedBuffer;
-        }
-
-        public Span<byte> GetSpan(int sizeHint = 0)
-        {
-            if (sizeHint > _rentedBuffer.Length)
-            {
-                var newSize = _rentedBuffer.Length * 2;
-                if (sizeHint != 0)
-                {
-                    newSize = sizeHint;
-                }
-                var temp = _rentedBuffer;
-                _rentedBuffer = ArrayPool<byte>.Shared.Rent(newSize);
-                temp.AsSpan().Clear();
-                ArrayPool<byte>.Shared.Return(temp);
-            }
-            Debug.Assert(_rentedBuffer.Length > 0);
-            return _rentedBuffer;
-        }
-    }
-
     public class DependencyContextWriter
     {
         public void Write(DependencyContext context, Stream stream)
@@ -114,6 +42,7 @@ namespace Microsoft.Extensions.DependencyModel
             {
                 WriteRuntimeGraph(context, ref jsonWriter);
             }
+            jsonWriter.Flush();
         }
 
         private void WriteRuntimeTargetInfo(DependencyContext context, ref Utf8JsonWriter jsonWriter)
